+++ nx2/blocks/chat/chat.css
@@ -213,7 +213,7 @@
       }
 
       &::before {
-        margin-top: 6px;
+        margin-top: var(--s2-spacing-75);
       }
     }
 
@@ -243,26 +243,27 @@
     width: 14px;
     height: 14px;
     flex-shrink: 0;
-    margin-top: 4px;
+    margin-top: var(--s2-spacing-75);
     appearance: none;
     border: 2px solid var(--s2-gray-800);
     border-radius: var(--s2-corner-radius-75);
     background: var(--s2-gray-75);
 
     &:disabled {
-      opacity: 0.3;
+      opacity: 0.8;
       cursor: not-allowed;
     }
 
     &:checked {
       background: var(--s2-gray-800);
+      color: var(--s2-gray-25);
 
       &::after {
         content: "";
         position: absolute;
         inset: 0;
-        background-color: var(--s2-gray-25);
-        mask-image: url("/nx2/img/icons/S2_Icon_Checkmark_20_N.svg");
+        background-color: currentcolor;
+        mask-image: url("https://da.live/img/icons/s2-icon-checkmark-20-n.svg");
         mask-size: contain;
         mask-repeat: no-repeat;
         mask-position: center;
