+++ nx2/blocks/chat/utils/persistence.js
